+++ src/date-range-picker/dropdown.tsx
@@ -8,7 +8,7 @@ import { ButtonProps } from '../button/interfaces';
 import { InternalButton } from '../button/internal';
 import FocusLock from '../internal/components/focus-lock';
 import InternalBox from '../box/internal';
-import SpaceBetween from '../space-between/index.js';
+import InternalSpaceBetween from '../space-between/internal';
 
 import styles from './styles.css.js';
 import RelativeRangePicker from './relative-range';
@@ -149,9 +149,9 @@ export function DateRangePickerDropdown({
               [styles['one-grid']]: isSingleGrid,
             })}
           >
-            <SpaceBetween size="l">
+            <InternalSpaceBetween size="l">
               <InternalBox padding={{ top: 'm', horizontal: 'l' }}>
-                <SpaceBetween direction="vertical" size="s">
+                <InternalSpaceBetween direction="vertical" size="s">
                   {rangeSelectorMode === 'default' && (
                     <ModeSwitcher
                       mode={rangeSelectionMode}
@@ -188,7 +188,7 @@ export function DateRangePickerDropdown({
                       i18nStrings={i18nStrings}
                     />
                   )}
-                </SpaceBetween>
+                </InternalSpaceBetween>
 
                 <InternalBox
                   className={styles['validation-section']}
@@ -224,7 +224,7 @@ export function DateRangePickerDropdown({
                   </div>
                 )}
                 <div className={styles['footer-button-wrapper']}>
-                  <SpaceBetween size="xs" direction="horizontal">
+                  <InternalSpaceBetween size="xs" direction="horizontal">
                     <InternalButton
                       onClick={closeDropdown}
                       className={styles['cancel-button']}
@@ -242,10 +242,10 @@ export function DateRangePickerDropdown({
                     >
                       {i18nStrings.applyButtonLabel}
                     </InternalButton>
-                  </SpaceBetween>
+                  </InternalSpaceBetween>
                 </div>
               </div>
-            </SpaceBetween>
+            </InternalSpaceBetween>
           </div>
         </div>
       </FocusLock>
